+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Ver. 0.5.0
 * :class:`Actor` internal message passing uses the (unmasked) websocket protocol
   in a bidirectional communication between the :class:`Arbiter` and actors.
 * :class:`CPUboundApplication` created as base class for CPU bound applications.
+* Added ability to add Websocket subprotocols and extensions.
 
 Ver. 0.4.4 - 2013-Jan-13
 ==============================